move welcome page styles to modules
