+++ src/index.css
@@ -25,7 +25,7 @@ body {
 }
 
 main {
-    padding: 140px 0;
+    padding-top: 140px;
     min-height: 90vh;
 }
