+++ src/pages/welcome-page/WelcomePage.module.css
@@ -1,0 +1,79 @@
+.welcomeContainer {
+    width: 100%;
+    margin-top: -2.5rem;
+    height: 90dvh;
+    background: var(--color-gray-50, #F9FAFB);
+    background: linear-gradient(205deg, #F9FAFB 40%, #A3B3FF 80%, rgba(79, 57,246, 0.5) 100%);
+}
+
+:global(.dark) .welcomeContainer {
+    background: #101828;
+    background: linear-gradient(205deg, #101828 50%, #312C85 90%, rgba(79, 57,246, 0.5) 100%);
+}
+
+.welcomeTextContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    min-height: 70dvh;
+}
+
+.welcomeHeader {
+    font-size: 3.75rem;
+    font-weight: 700;
+    color: var(--color-gray-900, #101828);
+}
+
+:global(.dark) .welcomeHeader {
+    color: var(--color-gray-50, #F9FAFB);
+}
+
+.welcomeLogo {
+    width: 300px;
+    margin: 0 auto 3rem auto;
+}
+
+.buttonContainer {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.button {
+    border-radius: 1rem;
+    min-width: 300px;
+    min-height: 40px;
+    font-size: 1.125rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: all 0.3s ease-in-out;
+}
+
+.loginButton {
+    color: var(--color-gray-50, #F9FAFB);
+    background-color: rgba(79, 57, 246, 0.9);
+}
+
+.loginButton:hover {
+    background-color: var(--color-indigo-600, #4F39F6);
+    box-shadow: 0 10px 15px -3px rgb(79, 57, 246, 0.5), 0 4px 6px -4px rgb(79, 57, 246, 0.5);
+}
+
+.guestButton {
+    color: var(--color-indigo-600, #4F39F6);
+    background: transparent;
+    border: 2px solid var(--color-indigo-600, #4F39F6);
+    cursor: pointer;
+}
+
+.guestButton:hover {
+    background-color: var(--color-indigo-50, #EEF2FF);
+}
+
+:global(.dark) .guestButton:hover {
+    background-color: rgba(49, 44, 133, 0.3);
+}
+
